Rename central pipeline from nodescanv2 to nodeinventory
