+++ central/sensor/service/pipeline/all/factory.go
@@ -19,8 +19,8 @@ import (
 	"github.com/stackrox/rox/central/sensor/service/pipeline/namespaces"
 	"github.com/stackrox/rox/central/sensor/service/pipeline/networkflowupdate"
 	"github.com/stackrox/rox/central/sensor/service/pipeline/networkpolicies"
+	"github.com/stackrox/rox/central/sensor/service/pipeline/nodeinventory"
 	"github.com/stackrox/rox/central/sensor/service/pipeline/nodes"
-	"github.com/stackrox/rox/central/sensor/service/pipeline/nodescansv2"
 	"github.com/stackrox/rox/central/sensor/service/pipeline/podevents"
 	"github.com/stackrox/rox/central/sensor/service/pipeline/processindicators"
 	"github.com/stackrox/rox/central/sensor/service/pipeline/reprocessing"
@@ -66,7 +66,7 @@ func (s *factoryImpl) PipelineForCluster(ctx context.Context, clusterID string) 
 		auditlogstateupdate.GetPipeline(),
 	}
 	if features.RHCOSNodeScanning.Enabled() {
-		pipelines = append(pipelines, nodescansv2.GetPipeline())
+		pipelines = append(pipelines, nodeinventory.GetPipeline())
 	}
 	if features.ComplianceOperatorCheckResults.Enabled() {
 		pipelines = append(pipelines,
