+++ central/sensor/service/pipeline/nodeinventory/pipeline.go
@@ -1,4 +1,4 @@
-package nodescansv2
+package nodeinventory
 
 import (
 	"context"
